chore: Remove FTP deployment workflow
The FTP deployment workflow in `.github/workflows/deploy.yml` has been removed. This is part of a move towards a different deployment strategy or potentially a discontinuation of automated deployments via FTP.

The `package-lock.json` file was also present in the diff, but no functional changes to dependencies are indicated. The primary change is the removal of the deployment workflow.